requirements.txt
httpx>=0.27,<1.0
pyyaml>=6.0,<7.0
pytest>=8.0,<9.0
